+++ class-wc-gateway-spectrocoin.php
@@ -180,30 +180,39 @@ class WC_Gateway_Spectrocoin extends WC_Payment_Gateway
 					</p>
 					<ul>
 						<li>
+							<span>1. </span>
 							<?php printf('<a href="%s" target="_blank">%s</a> %s', esc_url('https://auth.spectrocoin.com/signup'), __('Sign up', 'spectrocoin-accepting-bitcoin'), __('for a Spectroin Account.', 'spectrocoin-accepting-bitcoin')); ?>
 						</li>
 						<li>
+							<span>2. </span>
 							<?php printf('<a href="%s" target="_blank">%s</a> %s', esc_url('https://auth.spectrocoin.com/login'), __('Log in', 'spectrocoin-accepting-bitcoin'), __('to your Spectroin account.', 'spectrocoin-accepting-bitcoin')); ?>
 						</li>
 						<li>
+							<span>3. </span>
 							<?php printf('%s <b><a href="%s" target="_blank">%s</a></b> %s', __('On the dashboard, locate the', 'spectrocoin-accepting-bitcoin'), esc_url('https://spectrocoin.com/en/merchants/projects'), __('Business', 'spectrocoin-accepting-bitcoin'), __('tab and click on it.', 'spectrocoin-accepting-bitcoin')); ?>
 						</li>
 						<li>
+							<span>4. </span>
 							<?php printf('%s <b><a href="%s" target="_blank">%s</a>.</b>', __('Click on', 'spectrocoin-accepting-bitcoin'), esc_url('https://spectrocoin.com/en/merchants/projects/new'), __('New project', 'spectrocoin-accepting-bitcoin')); ?>
 						</li>
 						<li>
+							<span>5. </span>
 							<?php _e('Fill in the project details and select desired settings (settings can be changed).', 'spectrocoin-accepting-bitcoin'); ?>
 						</li>
 						<li>
+							<span>6. </span>
 							<?php printf('%s <b>%s</b> %s', __('The', 'spectrocoin-accepting-bitcoin'), __('Private Key', 'spectrocoin-accepting-bitcoin'), __('can be obtained by switching on the Public key radio button (Private key won\'t be visible in the settings window, and it will have to be regenerated in settings). Copy or download the newly generated private key.', 'spectrocoin-accepting-bitcoin')); ?>
 						</li>
 						<li>
+							<span>7. </span>
 							<?php _e('Click Submit.', 'spectrocoin-accepting-bitcoin'); ?>
 						</li>
 						<li>
+							<span>8. </span>
 							<?php _e('Copy and paste the Merchant ID and Project ID.', 'spectrocoin-accepting-bitcoin'); ?>
 						</li>
 						<li>
+							<span>9. </span>
 							<?php _e('Generate a test product. Create a test page on your WordPress website with a payment form connected to the Spectroin payment gateway. Perform a trial transaction using the test payment gateway (Test mode can be activated in project settings) to validate the integration\'s functionality. Verify the transaction details on the Spectroin dashboard to ensure it was successfully processed.', 'spectrocoin-accepting-bitcoin'); ?>
 						</li>
 						<br>